+++ commerce/warehouse/src/main/resources/schema.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS warehouse_product (
     product_id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    weight DOUBLE,
-    width DOUBLE,
-    height DOUBLE,
-    depth DOUBLE,
+    weight DOUBLE PRECISION,
+    width DOUBLE PRECISION,
+    height DOUBLE PRECISION,
+    depth DOUBLE PRECISION,
     fragile BOOLEAN,
     quantity INT
 );
@@ -11,8 +11,8 @@ CREATE TABLE IF NOT EXISTS warehouse_product (
 CREATE TABLE IF NOT EXISTS bookings (
     booking_id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     fragile BOOLEAN,
-    delivery_volume DOUBLE NOT NULL,
-    delivery_weight DOUBLE NOT NULL,
+    delivery_volume DOUBLE PRECISION NOT NULL,
+    delivery_weight DOUBLE PRECISION NOT NULL,
     delivery_id UUID,
     order_id UUID NOT NULL
 );
